Showed new filters, merged aggregations and additionalAggregations arrays, SizeFilter2 renamed to SizeFilter, deleted old SizeFilter
