+++ src/v2/Apps/Artist/Routes/WorksForSale/Utils/getWorksForSaleRouteVariables.ts
@@ -1,5 +1,4 @@
 import { paramsToCamelCase } from "v2/Components/ArtworkFilter/Utils/urlBuilder"
-import { getENV } from "v2/Utils/getENV"
 import { initialArtworkFilterState } from "v2/Components/ArtworkFilter/ArtworkFilterContext"
 import { allowedFilters } from "v2/Components/ArtworkFilter/Utils/allowedFilters"
 
@@ -13,15 +12,18 @@ export function getWorksForSaleRouteVariables({ artistID }, props) {
     ...paramsToCamelCase(filterStateFromUrl),
   }
 
-  const aggregations = ["MEDIUM", "TOTAL", "MAJOR_PERIOD"]
-  const additionalAggregations = getENV("ENABLE_NEW_ARTWORK_FILTERS")
-    ? ["PARTNER", "LOCATION_CITY", "MATERIALS_TERMS"]
-    : ["GALLERY", "INSTITUTION"]
-  const allAggregations = aggregations.concat(additionalAggregations)
+  const aggregations = [
+    "MEDIUM",
+    "TOTAL",
+    "MAJOR_PERIOD",
+    "PARTNER",
+    "LOCATION_CITY",
+    "MATERIALS_TERMS",
+  ]
 
   return {
     input: allowedFilters(filterParams),
-    aggregations: allAggregations,
+    aggregations,
     artistID,
   }
 }
